feat(auth-server): password encryption w/ bcrypt + auth-server code refactor
