+++ apps/auth-server/src/lib/util.ts
@@ -1,0 +1,9 @@
+export function exclude<User, Key extends keyof User>(
+  user: User,
+  keys: Key[]
+): Omit<User, Key> {
+  for (let key of keys) {
+    delete user[key];
+  }
+  return user;
+}
